Fix type casting for location price check

+++ src/GuildHouseBroker.cpp
@@ -127,7 +127,7 @@ bool GuildHouseBroker::OnGossipSelect(Player* player, Creature* creature, uint32
             return true;
         }
     
-        if (!player->HasEnoughMoney(location->Price))
+        if (!player->HasEnoughMoney(uint32(location->Price)))
         {
             ChatHandler(player->GetSession()).PSendSysMessage("Your guild does not have enough gold.");
             CloseGossipMenuFor(player);
